Traktor: PS3 updates. Mostly static analyze things from SN compiler fixes which is always a good thing to fix. Also able to build PS3 using VS2012.
[git-p4: depot-paths = "//depot/traktor/": change = 18096]

+++ code/Online/Psn/PsnStatistics.cpp
@@ -7,12 +7,12 @@ namespace traktor
 
 T_IMPLEMENT_RTTI_CLASS(L"traktor.online.PsnStatistics", PsnStatistics, IStatisticsProvider)
 
-bool PsnStatistics::enumerate(std::map< std::wstring, float >& outStats)
+bool PsnStatistics::enumerate(std::map< std::wstring, int32_t >& outStats)
 {
 	return true;
 }
 
-bool PsnStatistics::set(const std::wstring& statId, float value)
+bool PsnStatistics::set(const std::wstring& statId, int32_t value)
 {
 	return false;
 }
